+++ src/views/payment-sheet-views/card-view.js
@@ -159,6 +159,8 @@ CardView.prototype._generateHostedFieldsOptions = function () {
         return;
       }
 
+      normalizeStyles(overrides.styles[style]);
+
       assign(options.styles[style], overrides.styles[style]);
     });
   }
@@ -284,10 +286,10 @@ CardView.prototype.tokenize = function () {
 
 CardView.prototype.showFieldError = function (field, errorMessage) {
   var fieldError;
-  var fieldGroup = this.getElementById(camelCaseToSnakeCase(field) + '-field-group');
+  var fieldGroup = this.getElementById(camelCaseToKebabCase(field) + '-field-group');
 
   if (!this.fieldErrors.hasOwnProperty(field)) {
-    this.fieldErrors[field] = this.getElementById(camelCaseToSnakeCase(field) + '-field-error');
+    this.fieldErrors[field] = this.getElementById(camelCaseToKebabCase(field) + '-field-error');
   }
 
   classlist.add(fieldGroup, 'braintree-form__field-group--has-error');
@@ -303,10 +305,10 @@ CardView.prototype.showFieldError = function (field, errorMessage) {
 };
 
 CardView.prototype.hideFieldError = function (field) {
-  var fieldGroup = this.getElementById(camelCaseToSnakeCase(field) + '-field-group');
+  var fieldGroup = this.getElementById(camelCaseToKebabCase(field) + '-field-group');
 
   if (!this.fieldErrors.hasOwnProperty(field)) {
-    this.fieldErrors[field] = this.getElementById(camelCaseToSnakeCase(field) + '-field-error');
+    this.fieldErrors[field] = this.getElementById(camelCaseToKebabCase(field) + '-field-error');
   }
 
   classlist.remove(fieldGroup, 'braintree-form__field-group--has-error');
@@ -327,7 +329,7 @@ CardView.prototype._generateFieldSelector = function (field) {
 
 CardView.prototype._onBlurEvent = function (event) {
   var field = event.fields[event.emittedBy];
-  var fieldGroup = this.getElementById(camelCaseToSnakeCase(event.emittedBy) + '-field-group');
+  var fieldGroup = this.getElementById(camelCaseToKebabCase(event.emittedBy) + '-field-group');
   var activeId = document.activeElement && document.activeElement.id;
   var isHostedFieldsElement = document.activeElement instanceof HTMLIFrameElement && activeId.indexOf('braintree-hosted-field') !== -1;
 
@@ -381,7 +383,7 @@ CardView.prototype._onCardTypeChangeEvent = function (event) {
 };
 
 CardView.prototype._onFocusEvent = function (event) {
-  var fieldGroup = this.getElementById(camelCaseToSnakeCase(event.emittedBy) + '-field-group');
+  var fieldGroup = this.getElementById(camelCaseToKebabCase(event.emittedBy) + '-field-group');
 
   classlist.add(fieldGroup, 'braintree-form__field-group--is-focused');
 };
@@ -446,12 +448,20 @@ CardView.prototype._isCardTypeSupported = function (cardType) {
   return supportedCardTypes.indexOf(configurationCardType) !== -1;
 };
 
-function camelCaseToSnakeCase(string) {
+function camelCaseToKebabCase(string) {
   return string.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase();
 }
 
 function capitalize(string) {
   return string[0].toUpperCase() + string.substr(1);
+}
+
+function normalizeStyles(styles) {
+  Object.keys(styles).forEach(function (style) {
+    var transformedKeyName = camelCaseToKebabCase(style);
+
+    styles[transformedKeyName] = styles[style];
+  });
 }
 
 module.exports = CardView;